+++ crates/full-node/sov-ethereum/src/signer.rs
@@ -3,6 +3,7 @@ use jsonrpsee::RpcModule;
 use reth_primitives::{TxKind, U256};
 use reth_rpc_eth_types::EthApiError;
 use sov_evm::{EthereumAuthenticator, Evm, RlpEvmTransaction};
+use sov_modules_api::capabilities::{Kernel, KernelWithSlotMapping};
 use sov_modules_api::macros::config_value;
 use sov_modules_api::ApiStateAccessor;
 use sov_rollup_interface::node::da::DaService;
@@ -13,10 +14,11 @@ const DEFAULT_CHAIN_ID: u64 = config_value!("CHAIN_ID");
 
 pub(crate) fn register_signer_rpc_methods<
     S: sov_modules_api::Spec,
+    K: Kernel<S::Storage> + KernelWithSlotMapping<S>,
     Da: DaService,
     RT: EthereumAuthenticator<S> + Send + Sync + 'static,
 >(
-    rpc: &mut RpcModule<Ethereum<S, Da, RT>>,
+    rpc: &mut RpcModule<Ethereum<S, K, Da, RT>>,
 ) -> Result<(), jsonrpsee::core::client::Error> {
     rpc.register_async_method("eth_accounts", |_parameters, ethereum, _| async move {
         Ok::<_, ErrorObjectOwned>(ethereum.eth_signer.signers())
@@ -43,7 +45,7 @@ pub(crate) fn register_signer_rpc_methods<
             }
 
             let raw_evm_tx = {
-                let mut state = ApiStateAccessor::<S>::new(ethereum.storage.borrow().clone());
+                let mut state = ethereum.api_state_accessor();
 
                 // set nonce if none
                 if transaction_request.nonce.is_none() {