+++ crates/module-system/sov-modules-rollup-blueprint/src/endpoints.rs
@@ -2,19 +2,19 @@ use sov_db::ledger_db::LedgerDb;
 use sov_ledger_apis::LedgerRoutes;
 use sov_modules_api::execution_mode::ExecutionMode;
 use sov_modules_api::hooks::ApplyBatchHooks;
+use sov_modules_api::rest::StorageReceiver;
 use sov_modules_api::{RuntimeEventProcessor, Spec};
 use sov_modules_stf_blueprint::{Runtime as RuntimeTrait, RuntimeEndpoints, TxReceiptContents};
 use sov_rollup_interface::zk::{ZkvmGuest, ZkvmHost};
 use sov_sequencer::batch_builders::standard::StdBatchBuilder;
 use sov_sequencer::batch_builders::BatchBuilder;
 use sov_sequencer::{BatchBuilderConfig, SequencerConfig, SequencerDb};
-use tokio::sync::watch;
 
 use crate::{FullNodeBlueprint, SequencerBlueprint};
 
 /// Register rollup's default RPC methods and Axum router.
 pub async fn register_endpoints<B, M>(
-    storage: watch::Receiver<<B::Spec as Spec>::Storage>,
+    storage: StorageReceiver<B::Spec>,
     ledger_db: &LedgerDb,
     sequencer_db: &SequencerDb,
     da_service: &B::DaService,
@@ -27,9 +27,8 @@ where
     <B::InnerZkvmHost as ZkvmHost>::Guest: ZkvmGuest<Verifier = <B::Spec as Spec>::InnerZkvm>,
     <B::OuterZkvmHost as ZkvmHost>::Guest: ZkvmGuest<Verifier = <B::Spec as Spec>::OuterZkvm>,
 {
-    let mut endpoints = B::Runtime::endpoints(storage.clone());
     let da_address = sequencer_config.da_address.clone();
-    let sequencer_router = match &sequencer_config.batch_builder {
+    let (api_state, sequencer_router) = match &sequencer_config.batch_builder {
         BatchBuilderConfig::Standard(bb_config) => {
             let batch_builder =
                 StdBatchBuilder::<(B::Spec, B::DaSpec, B::Runtime), B::Kernel>::create(
@@ -48,12 +47,17 @@ where
                 ledger_db.clone(),
             );
 
-            sequencer.rest_api_server("/sequencer")
+            (
+                sequencer.api_state(),
+                sequencer.rest_api_server("/sequencer"),
+            )
         }
         BatchBuilderConfig::Preferred => {
             todo!("Preferred sequencer is not yet supported")
         }
     };
+
+    let mut endpoints = B::Runtime::endpoints(api_state);
 
     // Sequencer endpoints.
     endpoints.axum_router = endpoints.axum_router.nest("/sequencer", sequencer_router);